--after can now also take an iteration count
Once we have `--after foo' meaning `up to but not including foo' and `--after +3foo' meaning `up to foo, and including foo, and two lines after', it's natural to extend the syntax further.  `--after *1foo' means `up to, but not including, one additional foo (i.e. the second foo)'.  The star and plus notations can be combined, so that `--after +3+*2*foo', `--after +3*2foo', `--after *2*+3+foo' and `--after *2+3foo' all match up to two lines after the third foo.

I chose `*' for much the same reasons as I earlier chose `+': it naturally indicates multiples of something (e.g. the third foo), and can't appear at the start of a regex, meaning that we don't need a delimiter to separate it from the regex.